Data.Heap.Indexed.Binomial

Description

Size-indexed binomial heaps.

Documentation

data Tree n a

A size-indexed binomial tree.

Constructors

Root a (Node n a)

data Node :: Nat -> * -> * where

A binomial tree, indexed by its size.

Constructors

NilN :: Node 0 a

(:<) :: !(Tree n a) -> Node n a -> Node (1 + n) a

data Binomial :: Nat -> Nat -> * -> * where

A size-indexed binomial heap.

The implementation is similar to:

Hinze, Ralf. “Functional Pearls: Explaining Binomial Heaps.” Journal of Functional Programming 9, no. 1 (January 1999): 93–104. doi:10.1017/S0956796899003317.

Wasserman, Louis. “Playing with Priority Queues.” The Monad.Reader, May 12, 2010.

However invariants are more aggressively maintained, using a typechecker plugin. It is a list of binomial trees, equivalent to a binary number (stored least-significant-bit first).

Constructors

Nil :: Binomial n 0 a

(:-) :: !(Tree z a) -> Binomial (1 + z) xs a -> Binomial z ((1 + xs) + xs) a infixr 5

Skip :: Binomial (1 + z) (1 + xs) a -> Binomial z ((2 + xs) + xs) a