Data.Heap.Indexed.Braun

Description

Type-indexed Braun heaps.

Documentation

data Braun n a where

A Braun heap. Somewhat based on this implementation, but with a different strategy for maintaining invariants.

A braun tree is one where every left branch has at most one more element than the right branch.

Constructors

Leaf :: Braun 0 a

Node :: !(Offset n m) -> a -> Braun n a -> Braun m a -> Braun ((n + m) + 1) a

data Offset n m where

The "singleton" for whether or not the left branch is larger than the right.

Constructors

Even :: Offset n n

Lean :: Offset (1 + n) n