Data.Heap.Indexed.Class

Description

Classes for common functions between the heaps.

Documentation

class IndexedPriorityQueue h a where

A classed for indexed priority queues. Equivalent to PriorityQueue except the queues are indexed by their sizes.

Minimal complete definition

insert, empty, minViewMay, minView

Methods

empty :: h 0 a

The empty queue

minView :: h (1 + n) a -> (a, h n a)

Return the minimal element, and the rest of the queue.

singleton :: a -> h 1 a

A queue with one element.

insert :: a -> h n a -> h (1 + n) a

Add an element to the queue.

minViewMay :: h n a -> (n ~ 0 => b) -> (forall m. (1 + m) ~ n => a -> h m a -> b) -> b

Pattern match on the queue, and provide a proof that it is/isn't empty to the caller.

class IndexedPriorityQueue h a => MeldableIndexedQueue h a where

Queues which can be merged. Conforming members should form a monoid under merge and empty.

Minimal complete definition

merge

Methods

merge :: h n a -> h m a -> h (n + m) a

Merge two heaps. This operation is associative, and has the identity of empty.

merge x (merge y z) = merge (merge x y) z

merge x empty = merge empty x = x