Data.RoseTree

Documentation

data Tree a

A simple rose tree for use in some of the heaps.

Constructors

Root a [Tree a]

replicateA :: Applicative f => Int -> f a -> f (Tree a)

replicateA n x replicates the action x n times.

replicateTree :: Int -> a -> Tree a

foldTree :: (a -> [b] -> b) -> Tree a -> b