Data.Traversable.Sort

Description

Sort any traversable. Idea from here, but parameterized over the heap type.

Documentation

data Sort f a r where

A heap with a certain number of elements, and a function which extracts exactly that many elements from a larger heap.

Constructors

Sort :: (forall n. f (m + n) a -> (f n a, r)) -> !(f m a) -> Sort f a r

liftSort :: IndexedPriorityQueue f x => x -> Sort f x x

Lift a value into the running sort.

sortTraversable :: (MeldableIndexedQueue f a, Traversable t) => p f -> t a -> t a

Sort an arbitrary traversable container using a heap. The first parameter is a phantom, specifying which heap to use.

runSort :: forall x a f. Sort f x a -> a

Run the built-up function on the stored heap.

sortTraversal :: IndexedPriorityQueue f a => ((a -> Sort f a a) -> t -> Sort f a t) -> t -> t

Sort a traversal.