Data.Tree.Replicate

Description

Functions for creating rose trees (from Data.Tree) of a specified size.

Documentation

replicateA :: Applicative f => Int -> f a -> f (Tree a)

replicateA n x replicates the action x n times.

replicateTree :: Int -> a -> Tree a

replicateTree n a creates a tree of size n filled with a.

>>> putStr (drawTree (replicateTree 4 "."))
.
|
+- .
|
`- .
   |
   `- .

n > 0 ==> length (replicateTree n x) == n