TypeLevel.Singletons

Description

Provides singletons and general type-level utilities. singletons are value-level representations of types.

Eisenberg, Richard A., and Stephanie Weirich. “Dependently Typed Programming with Singletons.” In Proceedings of the 2012 Haskell Symposium, 117–130. Haskell ’12. New York, NY, USA: ACM, 2012. doi:10.1145/2364506.2364522.

Documentation

data family The k :: k -> *

A data family for singletons. The cute name allows code like this:

addZeroZero :: The Nat n -> n + 0 :~: n

class KnownSing x where

Class for singletons which can be generated.

Minimal complete definition

sing

Methods

sing :: The k x

(*.) :: The Nat n -> The Nat m -> The Nat (n * m) infixl 7

Multiply two numbers, on both the value and type level.

(+.) :: The Nat n -> The Nat m -> The Nat (n + m) infixl 6

Add two numbers, on both the value and type level.

(^.) :: The Nat n -> The Nat m -> The Nat (n ^ m) infixr 8

Raise a number to a power, on the type-level and value-level.

(<=.) :: The Nat n -> The Nat m -> The Bool (n <=? m) infix 4

Test order between two numbers, and provide a proof of that order with the result.

totalOrder :: p n -> q m -> ((n <=? m) :~: False) -> (m <=? n) :~: True

A proof of a total order on the naturals.

type (<=) x y = (x <=? y) :~: True

A proof that x is less than or equal to y.

type family (a :: Nat) + (b :: Nat) :: Nat where ... infixl 6

Addition of type-level naturals.

type family (a :: Nat) * (b :: Nat) :: Nat where ... infixl 7

Multiplication of type-level naturals.

type family (a :: Nat) ^ (b :: Nat) :: Nat where ... infixr 8

Exponentiation of type-level naturals.

data Nat :: *

(Kind) This is the kind of type-level natural numbers.