Data.Traversable.Sort

Description

Sort any traversable. Idea from here, but parameterized over the heap type.

Documentation

data Sort f a b r where

A heap with a certain number of elements, and a function which extracts exactly that many elements from a larger heap.

Constructors

Sort :: (forall n. f (m + n) b -> (f n b, r)) -> !(f m a) -> Sort f a b r

liftSort :: (IndexedQueue f a, IndexedQueue f x) => x -> Sort f x a a

Lift a value into the running sort.

sortTraversable :: (MeldableIndexedQueue f a, Traversable t) => p f -> t a -> t a

Sort an arbitrary traversable container using a heap. The first parameter is a phantom, specifying which heap to use.

runSort :: forall a b f. Sort f b b a -> a

Run the built-up function on the stored heap.

sortTraversal :: IndexedQueue f a => ((a -> Sort f a a a) -> t -> Sort f a a t) -> t -> t

Sort a traversal.

runSortWith :: forall a b c f. (forall n. f n a -> f n b) -> Sort f a b c -> c

transformTraversable :: (MeldableIndexedQueue f a, IndexedQueue f b, Traversable t) => (forall n. f n a -> f n b) -> t a -> t b